Fix handling of '-' token in BRDA lines.
The LCov spec says that '-' is a valid fourth token where the
block containing the branch was never executed and as such, the
branch could never have been taken. This is subtly different from
'branch never taken'.

lcov-result-merger did not handle this case (it just treated it
as a number) and the result was that branch execution counts could
be 'NaN'.